Extract activate/deactivate logic to a hook
